requirements.txt
hypercorn==0.17.3
python-telegram-bot[rate-limiter]==21.6
aiosqlite==0.20.0
Flask>=3.0,<4.0